+++ services/api/src/Firefly.Signal.JobSearch.Api/Contracts/Requests/ExportJobsRequest.cs
@@ -1,0 +1,4 @@
+namespace Firefly.Signal.JobSearch.Contracts.Requests;
+
+public sealed record ExportJobsRequest(
+    IReadOnlyList<long>? JobIds = null);
