+++ services/api/src/Firefly.Signal.JobSearch.Api/Contracts/Requests/ImportJobsFromProviderRequest.cs
@@ -1,0 +1,16 @@
+using Firefly.Signal.JobSearch.Application;
+
+namespace Firefly.Signal.JobSearch.Contracts.Requests;
+
+public sealed record ImportJobsFromProviderRequest(
+    int PageIndex = 1,
+    int PageSize = 50,
+    string Where = "london",
+    string? Keyword = null,
+    int DistanceKilometers = 5,
+    int MaxDaysOld = 30,
+    string Category = "it-jobs",
+    JobSearchProviderKind Provider = JobSearchProviderKind.Adzuna,
+    string? ExcludedKeyword = null,
+    decimal? SalaryMin = null,
+    decimal? SalaryMax = null);
