+++ services/api/src/Firefly.Signal.JobSearch.Api/Contracts/Requests/UpdateJobRequest.cs
@@ -1,0 +1,38 @@
+namespace Firefly.Signal.JobSearch.Contracts.Requests;
+
+public sealed record UpdateJobRequest(
+    long? JobRefreshRunId,
+    string SourceName,
+    string SourceJobId,
+    string? SourceAdReference,
+    string Title,
+    string Description,
+    string Summary,
+    string Url,
+    string Company,
+    string? CompanyDisplayName,
+    string? CompanyCanonicalName,
+    string Postcode,
+    string LocationName,
+    string? LocationDisplayName,
+    string? LocationAreaJson,
+    decimal? Latitude,
+    decimal? Longitude,
+    string? CategoryTag,
+    string? CategoryLabel,
+    decimal? SalaryMin,
+    decimal? SalaryMax,
+    string? SalaryCurrency,
+    bool? SalaryIsPredicted,
+    string? ContractTime,
+    string? ContractType,
+    bool IsFullTime,
+    bool IsPartTime,
+    bool IsPermanent,
+    bool IsContract,
+    bool IsRemote,
+    DateTime PostedAtUtc,
+    DateTime ImportedAtUtc,
+    DateTime LastSeenAtUtc,
+    bool IsHidden,
+    string RawPayloadJson);
